mc6883: tweak some types
Note uint8 compatible with range of previous unsigned for S in
serialisation.

+++ src/mc6883.h
@@ -28,7 +28,7 @@
 struct MC6883 {
 	struct part part;
 
-	unsigned S;
+	uint8_t S;
 	unsigned Zrow;
 	unsigned Zcol;
 	unsigned Vrow;
